Create usecase-utils (so we have one place to store all usecase/detail relevant functions)
