+++ webofneeds/won-owner-webapp/src/main/webapp/app/service/need-builder.js
@@ -6,7 +6,7 @@
 //TODO switch to requirejs for dependency mngmt (so this lib isn't angular-bound)
 //TODO replace calls to `won` object to `require('util')`
 import won from "./won.js";
-import { getAllDetails } from "../won-utils";
+import * as useCaseUtils from "../usecase-utils";
 import { is } from "../utils";
 import { useCases } from "../../config/usecase-definitions.js";
 import {
@@ -83,7 +83,7 @@ import { Generator } from "sparqljs";
 
     const addContent = (contentNode, contentData) => {
       //TODO: CANT HANDLE "@id" details yet (see won-message-utils.js buildChatMessage(..) additionalContent part
-      const detailList = getAllDetails();
+      const detailList = useCaseUtils.getAllDetails();
 
       for (const detail of Object.values(detailList)) {
         // const detail = detailList[detailName];
