+++ webofneeds/won-owner-webapp/src/main/webapp/app/won-utils.js
@@ -19,14 +19,13 @@ import {
 
 import { ownerBaseUrl } from "config";
 import urljoin from "url-join";
-import * as useCaseDefinitions from "useCaseDefinitions";
-import { messageDetails } from "detailDefinitions";
 import qr from "qr-image";
 import jsonld from "jsonld";
 window.jsonld4dbg = jsonld;
 
 import Immutable from "immutable";
 import { get, parseDatetimeStrictly, isValidDate } from "./utils.js";
+import * as useCaseUtils from "./usecase-utils.js";
 
 import won from "./won-es6.js";
 
@@ -426,153 +425,6 @@ export function getRandomWonId() {
     getRandomString(11, "abcdefghijklmnopqrstuvwxyz0123456789")
   );
 }
-/**
- * Returns all the details that are defined in any useCase Defined in the useCaseDefinitions
- * and in the messageDetails
- */
-export function getAllDetails() {
-  let allDetails = {};
-  const useCases = useCaseDefinitions.getUseCases();
-
-  if (hasSubElements(useCases)) {
-    for (const useCaseKey in useCases) {
-      const useCase = useCases[useCaseKey];
-      if (useCase) {
-        const details = useCase.details ? useCase.details : {};
-        const seeksDetails = useCase.seeksDetails ? useCase.seeksDetails : {};
-        allDetails = { ...allDetails, ...details, ...seeksDetails };
-      }
-    }
-  }
-
-  return Object.assign({}, messageDetails, allDetails);
-}
-
-/**
- * Returns the corresponding UseCase to a given need,
- * return undefined if no useCase is found,
- * @param needImm
- */
-export function findUseCaseByNeed(needImm) {
-  const seeksTypes =
-    getIn(needImm, ["seeks", "type"]) &&
-    getIn(needImm, ["seeks", "type"])
-      .toSet()
-      .remove("won:Need");
-
-  const contentTypes =
-    getIn(needImm, ["content", "type"]) &&
-    getIn(needImm, ["content", "type"])
-      .toSet()
-      .remove("won:Need");
-  const useCases = useCaseDefinitions.getUseCases();
-
-  if (hasSubElements(useCases)) {
-    const useCasesImm = Immutable.fromJS(useCases);
-
-    if (useCasesImm && useCasesImm.size > 0) {
-      const hasExactMatchingTypes = (useCase, types, branch) => {
-        const typesSize = types ? types.size : 0;
-        const ucTypes =
-          useCase.getIn(["draft", branch, "type"]) &&
-          useCase
-            .getIn(["draft", branch, "type"])
-            .toSet()
-            .remove("won:Need");
-
-        const ucTypesSize = ucTypes ? ucTypes.size : 0;
-
-        if (typesSize != ucTypesSize) {
-          return false;
-        }
-
-        const hasTypes = typesSize > 0;
-
-        return !hasTypes || (ucTypes && ucTypes.isSubset(types));
-      };
-
-      let matchingUseCases = useCasesImm.filter(useCase =>
-        hasExactMatchingTypes(useCase, contentTypes, "content")
-      );
-
-      if (matchingUseCases && matchingUseCases.size > 1) {
-        //If there are multiple matched found based on the content type(s) alone we refine based on the seeks type as well
-        matchingUseCases = matchingUseCases.filter(useCase =>
-          hasExactMatchingTypes(useCase, seeksTypes, "seeks")
-        );
-      }
-
-      if (matchingUseCases && matchingUseCases.size > 1) {
-        console.warn(
-          "Found multiple matching UseCases for: ",
-          needImm,
-          " matching UseCases: ",
-          matchingUseCases,
-          " -> returning undefined"
-        );
-        return undefined;
-      } else if (matchingUseCases && matchingUseCases.size == 1) {
-        return matchingUseCases.first().toJS();
-      } else {
-        console.warn(
-          "Found no matching UseCase for:",
-          needImm,
-          " within, ",
-          useCases,
-          " -> returning undefined"
-        );
-        return undefined;
-      }
-    }
-  }
-
-  return undefined;
-}
-
-window.findUseCaseByNeed4Dbg = findUseCaseByNeed;
-/**
- * Returns all the details that are defined in any useCase in the useCaseDefinitions
- * and has the messageEnabled Flag set to true
- *
- * as well as every detail that is defined in the messageDetail object from detail-definitions (for details that are only available in
- * messages)
- *
- * the messageEnabled-flag indicates if the detail is allowed to be sent as a part of a connectionMessage
- * @returns {{}}
- */
-export function getAllMessageDetails() {
-  let allDetails = {};
-  const useCases = useCaseDefinitions.getUseCases();
-
-  if (hasSubElements(useCases)) {
-    for (const useCaseKey in useCases) {
-      const useCase = useCases[useCaseKey];
-      if (useCase) {
-        const details = useCase.details ? useCase.details : {};
-        const seeksDetails = useCase.seeksDetails ? useCase.seeksDetails : {};
-        allDetails = { ...allDetails, ...details, ...seeksDetails };
-      }
-    }
-  }
-
-  let usecaseMessageDetails = {};
-
-  for (const detailKey in allDetails) {
-    if (allDetails[detailKey].messageEnabled) {
-      usecaseMessageDetails[detailKey] = allDetails[detailKey];
-    }
-  }
-  const allMessageDetails = Object.assign(
-    {},
-    messageDetails,
-    usecaseMessageDetails
-  );
-  return allMessageDetails;
-}
-
-function hasSubElements(obj) {
-  return obj && obj !== {} && Object.keys(obj).length > 0;
-}
 
 export function findLatestIntervallEndInJsonLdOrNowAndAddMillis(
   draft,
@@ -675,7 +527,7 @@ export function createDocumentDefinitionFromPost(post) {
   let content = [];
   content.push(title);
 
-  const allDetails = getAllDetails();
+  const allDetails = useCaseUtils.getAllDetails();
 
   const postContent = post.get("content");
   if (postContent) {
